+++ src/pages/cart/CartContent/EmptyCartContainer/EmptyCartContainer.styled.ts
@@ -1,0 +1,25 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: #0a0d13;
+  font-size: 16px;
+  font-weight: 400;
+`;
+
+export const PayConfirmButton = styled.button`
+  width: 500px;
+  height: 64px;
+  position: fixed;
+  bottom: 0;
+  left: 50%;
+  transform: translateX(-50%);
+  background-color: #bebebe;
+  color: #fff;
+  font-size: 16px;
+  font-weight: 700;
+`;
