+++ src/pages/cart/CartPage.tsx
@@ -1,6 +1,6 @@
 import * as S from './CartPage.styled';
 import Header from '@/shared/components/Header/Header';
-import CartContent from './CartContent';
+import CartContent from './CartContent/CartContent';
 
 export default function CartPage() {
   return (
