--- src/pages/cart/EmptyCartContainer/EmptyCartContainer.styled.ts
@@ -1,12 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Container = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: #0a0d13;
-  font-size: 16px;
-  font-weight: 400;
-`;
