--- src/pages/cart/EmptyCartContainer/EmptyCartContainer.tsx
@@ -1,5 +1,0 @@
-import * as S from './EmptyCartContainer.styled';
-
-export default function EmptyCartContainer() {
-  return <S.Container>장바구니에 담은 상품이 없습니다.</S.Container>;
-}
